pyproject.toml
[project]
name = "qa-trainings"
version = "0.1.0"
description = ""
authors = [
    {name = "TKM-h.almughrabi-c",email = "h.almughrabi-c@takamol.sa"}
]
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "flask (>=3.1.2,<4.0.0)",
    "flask-sqlalchemy (>=3.1.1,<4.0.0)",
    "pymysql (>=1.1.0,<2.0.0)",
    "pandas (>=2.3.3,<3.0.0)",
    "openpyxl (>=3.1.5,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
